fix one more circ dep
